raspberry-pi/requirements.txt
dlib
face-recognition
numpy
Pillow
gTTS
azure-storage-blob
azure-iot-device
configparser
